inventory-provisioners/esxi/requirements.txt
pyvmomi==7.0
PyYAML==5.3.1